Light : Defer to Session for creation/destruction
This will give us a place where we can hook in portal/dome management.

I think this pattern might work for cameras too. And if all destruction went through the Session, then that's where the "no delete for batch" thing could be implemented. And perhaps we could also do "no delete during ipr shutdown" if that made things quicker (it would if we could do it for Arnold).